spyder

+++ bai13-3.py
@@ -4,7 +4,7 @@ a = float(input("nhập dung lượng giới hạn(1-1024MB): "))*1024
 while ((a/1024)<1) or ((a/1024)>1024):
     a = float(input("yêu cầu nhập lại (1-1024MB)"))*1024
 n = int(a//1000) #số file
-c = (a%1000)*1024 #dung lượng file cuối(byte)
+c = int((a%1000)*1024) #dung lượng file cuối(byte)
 os.mkdir("thư mục chứa file")
 for i in range(n):
     os.chdir(add+"\\thư mục chứa file")
